Applies Compensation using specifications in csv file provided at path_comp_mat

The csv file at path_comp_mat should specify the channels to apply the compensation to. The format is a matrix where the col and row names correspond to the channel names

Usage

getCompGS(gs, path_comp_mat)

Arguments

gs

GatingSet to apply Biexponential Transformation to

path_comp_mat

file path for .csv file that specifies the Compensation Matrix

Value

GatingSet with compensated data

Details

An example is provided in the extdata/comp_mat_example_fcs.csv

Examples

if (FALSE) {
path_comp_mat = system.file("extdata", "comp_mat_example_fcs.csv",
                             package = "staRgate", mustWork = TRUE)

# gs is a GatingSet object
gs = getCompGS(gs, path_comp_mat = path_comp_mat)
}