Attach indicator columns to intens_dat based on gates provided in cutoffs

Adds an indicator column (0/1) to intens_dat for each marker in cutoffs as indicated by the columns in cutoffs

Usage

getGatedDat(intens_dat = intensity_dat, cutoffs, subset_col)

Arguments

intens_dat

dataframe of pre-gated (compensated, biexp. transf, openCyto steps) intensity values where rows = each sample

cutoffs

tibble of gates/cutoffs for all markers to gate Expects cutoffs to match format of output from getDensityGates() with column corresponding to a marker, and rows to the subsets defined in the subset_col

subset_col

string for the column name to indicate the subsets to apply density gating on will perform operation on subsets corresponding to each unique value in column

Value

intens_dat with additional cols attached for each marker in cutoffs

Details

The naming convention for the tagged on indicator columns will be tolower(<marker_name>_pos) where 0 indicates negativity or intensity < gate provided 1 indicates positivity or intensity > gate provided